refactor: fix comments on pull request
