+++ declarations.d.ts
@@ -4,12 +4,12 @@ declare module '*.gif';
 declare module '*.svg';
 
 export const enum PageIds {
-    adminPage = 'admin-page',
-    audiocallPage = 'audiocall-page',
-    errorPage = 'error-page',
-    mainPage = 'main-page',
-    settingsPage = 'settings-page',
-    sprintPage = 'sprint-page',
-    statsPage = 'stats-page',
-    textbookPage = 'textbook-page',
+  adminPage = 'admin-page',
+  audiocallPage = 'audiocall-page',
+  errorPage = 'error-page',
+  mainPage = 'main-page',
+  settingsPage = 'settings-page',
+  sprintPage = 'sprint-page',
+  statsPage = 'stats-page',
+  textbookPage = 'textbook-page',
 }
