+++ index.html
@@ -7,6 +7,6 @@
     <link rel="stylesheet" href="./src/style.css">
     <title>RS lang</title>
 </head>
-<body>
-</body>
+    <body>
+    </body>
 </html>
