+++ src/Pages/Admin/admin.scss
@@ -1,11 +1,10 @@
 * {
   margin: 0;
   padding: 0;
-  box-sizing: border-box;  
+  box-sizing: border-box;
 }
 
-
-.admin-container{
+.admin-container {
   display: flex;
   align-items: center;
   justify-content: center;
@@ -13,7 +12,7 @@
   height: 70vh;
 }
 
-p{
+p {
   text-transform: uppercase;
   font-size: 30px;
 }
