+++ src/Pages/Error/error.scss
@@ -1,13 +1,12 @@
-
-.error-container{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin: 0 auto;
-    height: 70vh;
+.error-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin: 0 auto;
+  height: 70vh;
 }
 
-p{
-    text-transform: uppercase;
-    font-size: 30px;
-  }
+p {
+  text-transform: uppercase;
+  font-size: 30px;
+}
