+++ src/Pages/Main/main.scss
@@ -1,10 +1,10 @@
 * {
   margin: 0;
   padding: 0;
-  box-sizing: border-box;  
+  box-sizing: border-box;
 }
 
-.main-container{
+.main-container {
   display: flex;
   align-items: center;
   justify-content: center;
@@ -12,7 +12,7 @@
   height: 70vh;
 }
 
-p{
+p {
   position: relative;
   text-transform: uppercase;
   font-size: 30px;
