+++ src/components/Footer/footer.scss
@@ -7,8 +7,7 @@
     .footer__cont{
       width: 100%;
       display: flex;
-      justify-content: space-around;
-      // margin: auto ;
+      justify-content: space-around;      
       align-items: center;
     }
