+++ src/components/Page/index.ts
